Send customer info to first listener added (#576)
Fixes [CSDK-611](https://revenuecats.atlassian.net/browse/CSDK-611). 

This is the approach that stores the last customer info on the dart side
and sends it to the first listener added. This approach, while not being
the cleanest, wouldn't require API changes as I talked in the ticket.
Tests for this are added in a followup PR
https://github.com/RevenueCat/purchases-flutter/pull/580


[CSDK-611]:
https://revenuecats.atlassian.net/browse/CSDK-611?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ