Fix purchase tester reconfiguring the SDK (#722)
Purchase tester currently reconfigures the RC SDK on every InitialScreen
instantiation. Due to the way we do navigation, we instantiate that
screen after a purchase and possibly in some other situations. This PR
does the following changes:
- Makes it so we don't recreate the `InitialScreen` without need in
purchase tester.
- Moves the SDK initialization to the `main.dart` file so we ensure it's
executed only once (This is done both in purchase tester and
MagicWeather)