Fix flutter static analysis in pub.dev (#726)
Looks like we had a bunch of warnings in pub.dev due to freezed
annotations. This PR hopes to fix those. Most of the warnings were in
the form:
```
warning • The annotation 'JsonKey' can only be used on fields or getters • lib/models/offering_wrapper.dart:45:6 • invalid_annotation_target
```
This is an issue recognized by the `freezed` library maintainer and
their recommendation is to just disable the lint rule, which we already
do locally... https://github.com/rrousselGit/freezed/issues/488.
However, it seems that in pub.dev, they don't recognize our ignore rules
and just run the entire set of rules.

In this PR, I removed the unnecessary `@JsonKey` annotations in our
models, and left a few that were causing some changes in behavior and
added a explicit ignore comment in those.

Additionally, in this PR, I updated flutter_lints