Remove result from presentPaywall methods (#916)
Currently the `presentPaywall` and `presentPaywallIfNeeded` methods
return a boolean indicating whether a purchase happened in the paywall.
This API is confusing and not extensible and it wasn't properly working
in iOS, see: #886. This first PR removes the return value from the
method. In a future PR we will add proper result handling for the
paywalls. Until then, the clients won't know what happened when the
paywall was presented and will need to rely on the
`Purchases.getCustomerInfo()` method to respond appropriately.

This is a breaking change from the previous beta.