Add dependency_overrides to purchases_flutter_ui (#1007)
I wasn't sure what was going on in #995 but there was some issues
importing a newly added model. The problem is that we were depending on
version ^6.0.0 of purchases_flutter.

I made some changes to the pubspec.yaml, so now we use the local plugin.

According to [the docs](https://dart.dev/tools/pub/dependencies), "As a
result, if you publish a package to pub.dev, keep in mind that your
package's dependency overrides are ignored by all users of your
package.". So we should be fine to publish it like this.