Fix new Flutter version test issues (#1289)
Looks like in the last version of flutter, api tests started failing
with an error:
```
You are applying Flutter's app_plugin_loader Gradle plugin imperatively using the apply script method, 
which is not possible anymore. Migrate to applying Gradle plugins 
with the declarative plugins block: https://flutter.dev/to/flutter-gradle-plugin-apply
```

This PR updates the gradle files to new ones obtained from a new sample
app, so we are actually using `gradle.kts` and the latest syntax.
Considering it's only for the api-tests, this seems like a low-risk
change.