Remove freezed dev dependency (#1366)
This reverts some changes done in #1365 in order to support freezed and
we remove freezed so we don't have issues causing us to have to increase
our min sdk version. Note that this also means we can't regenerate
freezed models anymore, but we plan to move away from freezed moving
forward.

---------

Co-authored-by: David <david@macrofactorapp.com>