Remove freezed generated code in favor of manually written mappings (#1368)
This follows up the work in #1365 and #1366. But this ends up removing
freezed completely in favor of writing our own mapping implementation. A
few points to note:
- While I tried to minimize breaking changes, there are a couple that I
wasn't sure how to fix.
- I had to add the Equatable package to be able to keep an equals
implementation same way we had it with freezed. We could also write
these manually (equals + hashcode) but felt like a lot of boilerplate,
and this package seems simpler... Lmk if you think otherwise!