Add purchase_tester dependency to fix issues in latest version of flutter (#1435)
Starting on version 3.35.0 of flutter, ios and macos tests started
failing. Trying to debug, it seems that the Podfile.lock didn't include
PHC nor purchases-ios dependencies. Looking around, I saw that we were
adding these local path dependencies as overrides, but not as
dependencies themselves. Adding them there made the dependencies appear
again in the Podfile.lock and for it to work again.

This shouldn't affect customers and be related to how we are depending
on our SDKs from the purchase tester app.