Fix WASM parsing (#1434)
There were some parsing issues when using WASM that made it impossible
to obtain offerings and CustomerInfo related to parsing double numbers
that were meant to be integers. This fixes those by just rounding to the
nearest integer. This should be fine since these values are meant to be
integer originally.

This should help dealing with #1433 

Still investigating another issue that makes the purchase flow get stuck
loading on WASM which will be addressed in future PRs