Default to Swift Package Manager in Purchase Tester (instead of CocoaPods) (#1454)
This PR removes all the traces of CocoaPods from the Purchase Tester
example app of the repo. This way, purchases_flutter is added as a local
dependency via Swift Package Manager.

The Purchase Tester project is also ready to seamlessly work with
CocoaPods for development when needed. To do that, just run
```bash
flutter config --no-enable-swift-package-manager
```
at the beginning of the session.

To re-enable Swift Package Manager, run 
```bash
flutter config --enable-swift-package-manager
```

## Notes for reviewer
The changes in this PR are:
* Remove all traces of CocoaPods from the Purchase Tester example app
* Commit all file changes generated by flutter when doing `flutter pub
get` and `flutter build ios (macos)` when using SPM