Add `purchase` method with `PurchaseParams` object (#1461)
This adds a new class `PurchaseParams` that allows to specify the
different possible parameters that you can pass to the purchase methods.

This pattern allows to add more parameters to the purchase methods,
making it more flexible. No new functionality is added in this PR, just
deprecates the old APIs in favor of the new one.

- [ ] Update docs