Support Purchasing Packages as Add-Ons (#1545)
This PR allows developers using the Flutter SDK to purchase a Play Store
subscription with `Package` add-ons by adding `addOnPackages` to the
`PurchaseParams` constructors.

It also updates the purchase tester's add-ons screen to allow you to
select whether each add-on item is purchased as a `Package`

This PR is marked to go into the `addOns-dev` branch. Once approved and
merged, we'll create a second add-ons beta release from the `addOns-dev`
branch.